backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
boto3==1.34.0
python-multipart==0.0.6
python-dateutil==2.8.2
mangum==0.17.0
lambda-decorators==0.6.0
lambda-warmer-py==0.6.0
